vanakkam_akka/backend/requirements.txt
# === Core Framework ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# === Database ===
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1
aiosqlite==0.19.0

# === Authentication ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# === AI / ML ===
openai==1.6.1
langchain==0.0.354
langchain-community==0.0.10
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
indicnlp==0.92

# === Tamil NLP ===
googletrans==4.0.0rc1
indic-transliteration==2.3.44

# === PDF & Documents ===
reportlab==4.0.8
python-docx==1.1.0
Pillow==10.1.0

# === Firebase ===
firebase-admin==6.2.0

# === WebSockets & Real-time ===
websockets==12.0

# === Scheduling & Background Tasks ===
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# === Utilities ===
httpx==0.25.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
uuid==1.30

# === Dev & Testing ===
pytest==7.4.4
pytest-asyncio==0.23.3
